requirements.txt
amqp==5.1.1
asgiref==3.5.0
billiard==3.6.4.0
boto3==1.22.2
botocore==1.25.2
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
convert-numbers==0.4
Django==3.2
django-celery-beat==2.2.1
django-ckeditor==6.4.0
django-js-asset==2.0.0
django-rosetta==0.9.8
django-smartfields==1.1.3
django-storages==1.12.3
django-timezone-field==4.2.3
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
kavenegar==1.1.2
Khayyam==3.0.17
kombu==5.2.4
Pillow==9.1.0
polib==1.1.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pymemcache==3.5.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
